+++ nix/packages/livekit-protocol.nix
@@ -1,0 +1,45 @@
+{
+  lib,
+  buildPythonPackage,
+  fetchFromGitHub,
+  setuptools,
+  protobuf,
+  types-protobuf,
+}:
+
+buildPythonPackage rec {
+  pname = "livekit-protocol";
+  version = "1.0.2";
+  pyproject = true;
+
+  src = fetchFromGitHub {
+    owner = "livekit";
+    repo = "python-sdks";
+    tag = "protocol-v${version}";
+    hash = "sha256-opYF9BmZ8reuyo6IeBjdJC6YQGypnWGHLSY3phJ4cxQ=";
+    fetchSubmodules = true;
+  };
+
+  sourceRoot = "source/livekit-protocol";
+
+  build-system = [
+    setuptools
+  ];
+
+  dependencies = [
+    protobuf
+    types-protobuf
+  ];
+  
+  pythonImportsCheck = [
+    "livekit"
+  ];
+
+  meta = {
+    description = "LiveKit real-time and server SDKs for Python";
+    homepage = "https://github.com/livekit/python-sdks/";
+    license = lib.licenses.asl20;
+    maintainers = with lib.maintainers; [ soyouzpanda ];
+    platforms = lib.platforms.all;
+  };
+}